+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerLauncherCodeGeneratorTest.kt
@@ -91,6 +91,39 @@ internal class ServerLauncherCodeGeneratorTest {
     }
 
     @Test
+    fun `generate with cache`() {
+        val openAPI = createOpenAPI()
+        context.addService(Context.SERVICE_CACHE)
+
+        codegen.generate(openAPI, context)
+
+        // Launcher
+        val file = File("${context.outputDirectory}/src/main/kotlin/com/wutsi/test/Application.kt")
+        assertTrue(file.exists())
+
+        val text = file.readText()
+        assertEquals(
+            """
+                package com.wutsi.test
+
+                import kotlin.String
+                import kotlin.Unit
+                import org.springframework.boot.autoconfigure.SpringBootApplication
+                import org.springframework.cache.`annotation`.EnableCaching
+
+                @SpringBootApplication
+                @EnableCaching
+                public class Application
+
+                public fun main(vararg args: String): Unit {
+                  org.springframework.boot.runApplication<Application>(*args)
+                }
+            """.trimIndent(),
+            text.trimIndent()
+        )
+    }
+
+    @Test
     fun `generate - do not overwrite`() {
         val openAPI = createOpenAPI()