+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Run maven in quiet mode with Github Workflow
 - Add cache configuration in `pom.xml`, `application.yml`
+- Add `@EnableCaching` in `Application` launcher class
 
 ## [0.0.12] 2021-03-19
 ### Added